refactor(core): moved baseprimitive to utils
