+++ @navikt/core/react/src/primitives/stack/Stack.tsx
@@ -1,16 +1,22 @@
 import React, { type HTMLAttributes, forwardRef } from "react";
 import type { OverridableComponent } from "../../utils-external";
 import { omit } from "../../utils-external";
-import { Slot } from "../../utils/components/slot/Slot";
-import { cl } from "../../utils/helpers";
-import type { AsChildProps } from "../../utils/types";
 import {
   PRIMITIVE_PROPS,
   type PrimitiveProps,
   useBasePrimitiveProps,
-} from "../base/BasePrimitive";
-import { getResponsiveProps, getResponsiveValue } from "../utilities/css";
-import type { ResponsiveProp, SpacingScale } from "../utilities/types";
+} from "../../utils/components/base-primitive/BasePrimitive";
+import {
+  getResponsiveProps,
+  getResponsiveValue,
+} from "../../utils/components/base-primitive/helpers/css";
+import type {
+  ResponsiveProp,
+  SpacingScale,
+} from "../../utils/components/base-primitive/helpers/types";
+import { Slot } from "../../utils/components/slot/Slot";
+import { cl } from "../../utils/helpers";
+import type { AsChildProps } from "../../utils/types";
 
 export type StackProps = HTMLAttributes<HTMLDivElement> & {
   /**
