+++ @navikt/core/react/src/utils/components/base-primitive/BasePrimitive.stories.tsx
@@ -1,11 +1,11 @@
 import type { Meta, StoryObj } from "@storybook/react-vite";
 import React from "react";
-import { Slot } from "../../utils/components/slot/Slot";
-import { cl } from "../../utils/helpers";
-import { Box } from "../box";
-import { HGrid } from "../grid";
-import { Hide, Show } from "../responsive";
-import { HStack, VStack } from "../stack";
+import { Box } from "../../../primitives/box";
+import { HGrid } from "../../../primitives/grid";
+import { Hide, Show } from "../../../primitives/responsive";
+import { HStack, VStack } from "../../../primitives/stack";
+import { cl } from "../../helpers";
+import { Slot } from "../slot/Slot";
 import { type PrimitiveProps, useBasePrimitiveProps } from "./BasePrimitive";
 
 interface BasePrimitiveProps extends PrimitiveProps {
