+++ @navikt/core/react/src/utils/components/base-primitive/helpers/css.ts
