+++ @navikt/core/react/src/utils/components/base-primitive/helpers/types.ts
